+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Fixed
+
+- Properly handle `IEnumerable<T>` as values in a Dictionary (#2)
+
 ## [0.6.0] - 2023-09-02
 
 ### Added